requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
celery==5.3.6
cffi==1.17.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.39.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.11
django-countries==7.5.1
django-cors-headers==4.6.0
django-celery-beat==2.6.0
django-celery-email==3.0.0
# django-celery-results
django-filter==24.3
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
django-redis==5.4.0
djangorestframework==3.15.0
djangorestframework_simplejwt==5.4.0
djoser==2.2.2
drf-yasg==1.21.7
flower==2.0.1
google-generativeai==0.8.4
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.32
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.0.3
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
setuptools
sentry-sdk==2.22.0
plaid-python==29.1.0
# dwollav2==2.2.1 
# stripe==11.6.0